refactor: expose index entry point
