+++ test/test-utils/make-pubsub-instances.ts
@@ -1,5 +1,5 @@
 import { PubSub } from '@google-cloud/pubsub';
-import { PubSub as MockPubSub } from '../../src/mock-pubsub';
+import { PubSub as MockPubSub } from '../../src';
 
 export function makePubSubInstances({ projectId }: { projectId?: string }) {
   return [
